requirements.txt
crewai[tools]
langchain-google-genai
duckduckgo-search
python-dotenv
